requirements.txt
aio-pika==9.5.5
aiofiles==24.1.0
aioredis==2.0.1
aiormq==6.8.1
alembic==1.14.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt==4.0.0
billiard==4.2.1
cached-property==1.5.2
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.2
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.115.8
fastapi_keycloak==1.0.11
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.26.0
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
keycloak==3.1.5
kombu==5.4.2
lazy_loader==0.4
Mako==1.3.9
MarkupSafe==3.0.2
multidict==6.1.0
networkx==3.4.2
numpy==2.2.4
opencv-python==4.11.0.86
packaging==24.2
pamqp==3.3.0
passlib==1.7.4
pika==1.3.2
pillow==11.1.0
prompt_toolkit==3.0.50
propcache==0.3.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
rsa==4.9
scikit-image==0.25.2
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
tifffile==2025.3.13
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.18.3
